+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix issue with global contextual search
 - Add MS Teams Extraction notifications (see [Documentation\DataExtractions\ExtractionTeamsNotifications.md])
 - Add ability to use Cohort Catalogue Filters in an Extraction Configuration
-
+- Add DistinctByDestinationPKS distinction strategy to the DatasetExtractionSource Extraction pipeline component
+ 
 ## [8.4.4] - 2025-05-08 
 
 - Re-add AggregateGraph functionality